fix: server-render locale-specific html attributes
