+++ src/app/(ar)/ar/blog/page.tsx
@@ -1,0 +1,10 @@
+import { getRegularPosts } from "@/lib/markdown";
+import { getBlogIndexMetadata } from "@/lib/metadata";
+import BlogPageClient from "@/app/(en)/blog/BlogPageClient";
+
+export const metadata = getBlogIndexMetadata("ar", "/ar/blog");
+
+export default function ArBlogPage() {
+  const posts = getRegularPosts("ar");
+  return <BlogPageClient posts={posts} lang="ar" />;
+}
