+++ src/app/(ar)/ar/contact-us/page.tsx
@@ -1,0 +1,5 @@
+import ContactPage from "@/app/(en)/contact-us/ContactPageClient";
+
+export default function ArContactPage() {
+  return <ContactPage lang="ar" />;
+}
