+++ src/app/(ar)/ar/services/page.tsx
@@ -1,0 +1,5 @@
+import ServicesPage from "@/app/(en)/services/ServicesPageClient";
+
+export default function ArServicesPage() {
+  return <ServicesPage lang="ar" />;
+}
