+++ src/app/(ar)/ar/support-us/page.tsx
@@ -1,0 +1,5 @@
+import SupportPage from "@/app/(en)/support-us/SupportPageClient";
+
+export default function ArSupportPage() {
+  return <SupportPage lang="ar" />;
+}
