+++ src/app/(ar)/services/page.tsx
@@ -1,0 +1,5 @@
+import ServicesPage from "@/app/(en)/services/ServicesPageClient";
+
+export default function ServicesPageRoute() {
+  return <ServicesPage lang="ar" />;
+}
