+++ src/app/(ar)/support-us/page.tsx
@@ -1,0 +1,5 @@
+import SupportPage from "@/app/(en)/support-us/SupportPageClient";
+
+export default function SupportPageRoute() {
+  return <SupportPage lang="ar" />;
+}
