+++ src/app/(en-root)/en/blog/page.tsx
@@ -1,0 +1,11 @@
+import { getRegularPosts } from "@/lib/markdown";
+import { getBlogIndexMetadata } from "@/lib/metadata";
+import BlogPageClient from "@/app/(en)/blog/BlogPageClient";
+
+export const metadata = getBlogIndexMetadata("en", "/en/blog");
+
+export default function EnBlogPage() {
+  const posts = getRegularPosts("en");
+
+  return <BlogPageClient posts={posts} lang="en" />;
+}
