+++ src/app/(en-root)/en/contact-us/page.tsx
@@ -1,0 +1,5 @@
+import ContactPage from "@/app/(en)/contact-us/ContactPageClient";
+
+export default function EnContactPage() {
+  return <ContactPage lang="en" />;
+}
