+++ src/app/(en-root)/en/services/page.tsx
@@ -1,0 +1,5 @@
+import ServicesPage from "@/app/(en)/services/ServicesPageClient";
+
+export default function EnServicesPage() {
+  return <ServicesPage lang="en" />;
+}
