--- src/app/ar/contact-us/page.tsx
@@ -1,5 +1,0 @@
-import ContactPage from "../../(en)/contact-us/ContactPageClient";
-
-export default function ArContactPage() {
-  return <ContactPage lang="ar" />;
-}
