--- src/app/ar/services/page.tsx
@@ -1,5 +1,0 @@
-import ServicesPage from "../../(en)/services/ServicesPageClient";
-
-export default function ArServicesPage() {
-  return <ServicesPage lang="ar" />;
-}
