--- src/app/contact-us/page.tsx
@@ -1,5 +1,0 @@
-import ContactPage from "../(en)/contact-us/ContactPageClient";
-
-export default function ContactPageRoute() {
-  return <ContactPage lang="ar" />;
-}
