--- src/app/en/contact-us/page.tsx
@@ -1,5 +1,0 @@
-import ContactPage from "../../(en)/contact-us/ContactPageClient";
-
-export default function EnContactPage() {
-  return <ContactPage lang="en" />;
-}
