--- src/app/en/layout.tsx
@@ -1,12 +1,0 @@
-import type { Metadata } from "next";
-import { getBaseLocaleMetadata } from "@/lib/metadata";
-
-export const metadata: Metadata = getBaseLocaleMetadata("en");
-
-export default function EnglishLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return children;
-}
