--- src/app/en/services/page.tsx
@@ -1,5 +1,0 @@
-import ServicesPage from "../../(en)/services/ServicesPageClient";
-
-export default function EnServicesPage() {
-  return <ServicesPage lang="en" />;
-}
