--- src/app/en/support-us/page.tsx
@@ -1,5 +1,0 @@
-import SupportPage from "../../(en)/support-us/SupportPageClient";
-
-export default function EnSupportPage() {
-  return <SupportPage lang="en" />;
-}
