--- src/app/en/writing-contest/page.tsx
@@ -1,8 +1,0 @@
-import WritingContestPage from "../../(en)/writing-contest/WritingContestPageClient";
-import { getPostsByCategory } from "@/lib/markdown";
-
-export default function EnWritingContestPage() {
-  const posts = getPostsByCategory("en", "writing-contest");
-
-  return <WritingContestPage lang="en" posts={posts} />;
-}
