--- src/app/services/page.tsx
@@ -1,5 +1,0 @@
-import ServicesPage from "../(en)/services/ServicesPageClient";
-
-export default function ServicesPageRoute() {
-  return <ServicesPage lang="ar" />;
-}
