--- src/app/writing-contest/page.tsx
@@ -1,8 +1,0 @@
-import WritingContestPage from "../(en)/writing-contest/WritingContestPageClient";
-import { getPostsByCategory } from "@/lib/markdown";
-
-export default function WritingContestPageRoute() {
-  const posts = getPostsByCategory("ar", "writing-contest");
-
-  return <WritingContestPage lang="ar" posts={posts} />;
-}
